chore: refactor product domain structure
