+++ src/Products/Show/MainPanel.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import MediaQuery from 'react-responsive';
-import breakpoints from '../common/breakpoints';
+import breakpoints from '../../common/breakpoints';
 import PriceAndId from './PriceAndId';
 import ActionButtons from './ActionButtons';
 import ColorButton from './ColorButton';
