+++ src/Products/Show/Recommendations.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import RecommendationCard from './RecommendationCard';
-import Subheading from '../common/Subheading';
+import Subheading from '../../common/Subheading';
 
 const Recommendations = styled.section`
   padding: 0 .5rem;
