+++ src/Products/Show/index.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import styled from 'styled-components';
 import MediaQuery from 'react-responsive';
-import Accordion from '../common/Accordion';
-import breakpoints from '../common/breakpoints';
+import Accordion from '../../common/Accordion';
+import breakpoints from '../../common/breakpoints';
 import Showcase from './Showcase';
 import MainPanel from './MainPanel';
 import Gallery from './Gallery';
